Overhaul
There is an issue with redeeming previous transactions via the way Im sending and receiving the data from the back end. What needs to happen is to let the front end handle the BEEF decoding via Transaction.FromBEEF, This would allow the overlay service to send back beef, and txid to ensure that the data is correct.